Adds annot elements to generated XML, fixing #28.
This was a pretty tricky feature to add without causing major disruption. The problem I ran into was that running `self.device.get_result()` would nuke the annot objects from the page objects. Thus, before I ran that code, I needed to add annotations to the page. The way I did that is in the _add_annots() function, which pulls the annotation information off of `page.annots` and adds it to a LTLayout object, before it gets processed. This works pretty well, but:

  1. The <annot> elements that I add to the tree do not get sorted properly and just wind up at the end of the tree (I'm OK with this for now), and
  1. Height, width, and x/y attributes aren't provided anywhere I could find so I needed to generate them myself using the bbox. Intuitively this seems fine, but I don't know enough to be sure.